requirement.txt
python-snap7==2.0.2
pyinstaller==6.15.0